[server] Flip execution order of updateDescendantPermissions
Summary:
Right now, we first query for descendants, and then process them. After processing we have a list to query for the next iteration of the loop.

But with our recent changes, querying can yield descendants at multiple depths. We need to process one depth at a time, so that we can make sure we get the parent processed before the child.

It's easier for us to go by depth-by-depth starting with the processing step, because the processing step will yield a list of threads to query that are all at the same depth.

This may be more clear in the final product than in this depth.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1377