[lib] Temporarily reintroduce deprecated MEMBERSHIP permission constant
Summary: If we don't include `'membership'` in `assertThreadPermissions` then the server code crashes as soon as it encounters a permissions blob with membership permission. We need to clear the permissions out of the database before we can remove the constant entirely.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1380