[server] Don't delete inaccessible threads if they have child threads
Summary: See code comment for explanation. I observed a thread with a `parent_thread_id` that didn't exist in the `threads` table, and after debugging it a bit and talking to @atul (he had created the thread, but didn't remember deleting the parent), I concluded that the parent must've been deleted because it was inaccessible.

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Subscribers: KatPo, palys-swm, atul, Adrian

Differential Revision: https://phabricator.ashoat.com/D1387