[server] Add code to soft-launch-migration.js to fix threads with missing parents
Summary:
I waw testing `soft-launch-migration.js` and saw that it crashes when processing a particular thread. I investigated the issue and found that the thread had a `parent_thread_id` that didn't exist in the `threads` table.

The previous diff should address this issue and prevent it from occurring in the future.

This diff makes sure that `soft-launch-migration.js` doesn't crash, and instead fixes the database corruption before processing the thread.

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1388