[native] Add option to store chat input bar height in a context
Summary: I decided to use ref here to avoid rendering all the components which use the context every time `ChatInputBar` updates the height. On the other hand, including a `ref.current` property in context value might cause the memoized value to be outdated, so I decided to include a ref instead of a numeric height.

Test Plan: Tested with next diffs.

Reviewers: ashoat