Creating a `metadata-creator.js` in creators. This is where future metadata will be inserted into the `metadata` table.
Summary:
This `metadata-creator.js` in creators is where future metadata will be stored.

The first metadata to be stored in the `metadata` table is the `db_version` which is initialized to 0.
In the future, more metadata can be added here.

Test Plan: Ran the SQL query and it worked as intended. A `db_version` is inserted with value 0, to indicate that the `db_version` begins at 0. `migrations.js` will take care of incrementing the `db_version` as migrations are added.

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2627