[2/5][server] Use expired time instead of current
Summary: We no longer update focused time after every ping so we can't say that a row is current. We have a mechanism that deletes expired focused rows, so if a row is old enough to be deleted, we shouldn't use it.

Test Plan: Modify focused time in the db to be older than 4 hours, send a message to that thread and check if there were no focused users.

Reviewers: atul

Subscribers: boristopalov, karol-bisztyga, Adrian, ashoat

Differential Revision: https://phabricator.ashoat.com/D2737