[server] Do not check timestamp when determining focused threads
Summary: We do not need to check the timestamp because if there's a row for the given user / thread, we are sure that this row is recent. All the rows that are not valid are removed by `deleteActivityForViewerSession` e.g. when a disconnection occurs.

Test Plan: Open a thread and check in db if the row is there. Close the thread and check if the row was deleted.

Reviewers: atul