[server] Do not check timestamp posting message send
Summary: We do not need to check the timestamp because if there's a row for the given user / thread, we are sure that this row is recent. All the rows that are not valid are removed by `deleteActivityForViewerSession` e.g. when a disconnection occurs.

Test Plan: Added a console log in `postMessageSend` to check the rows that are returned from db query.
1. Open a thread as one user and send a message to it as another user - a row should have `focused_user` set.
2. Close the thread (once by going back and once by closing the app), send a message - `focused_user` should not be set.

Reviewers: atul