[server] Do not check time when deleting orphaned focused rows
Summary: We do not need to check the timestamp because if there's a row for the given user / thread, we are sure that this row is recent. All the rows that are not valid are removed by `deleteActivityForViewerSession` e.g. when a disconnection occurs.

Test Plan: Modified `cron.js` to run the cleanup more frequently and verified that `deleteOrphanedActivity` run without throwing

Reviewers: atul