[native] Add `/lib/flow-typed/` to `.flowconfig` libs list
Summary:
Added `/flow-typed` lib folder in native, like in `/server/.flowconfig`, because `../lib` is included when running `flow .` in `native` folder, but its types definitions are not included.
It caused problems with `lib` files when running when running flow in `native`, although in `lib` everything works.
Test Plan:
This fix solved the problem with missing `tcomb` types when adding `validation-utils` to `lib` folder.
Reviewers:
palys-swm
Subscribers: